libtool: restore recipe-specific application of 0008-libtool-Avoid-relinking-when-cross-compiling-its-poi.patch
It was pointed out that it
"potentially has side effects for native builds. It breaks the ability
to run files in the .libs directory which in the cross case we don't
care about but some natives did once used to do that. The risk is they
find host things instead."

(From OE-Core rev: 549122ba73cb6c4fd2cb05fcbcb1ebe1e425ac7f)

Signed-off-by: Alexander Kanavin <alex@linutronix.de>
Signed-off-by: Mathieu Dubois-Briand <mathieu.dubois-briand@bootlin.com>