autotools: log when we're removing aclocal.m4
Some mysterious autotools errors are because upstream has a custom
aclocal.m4 that we're deleting it unless we know we're not even running
aclocal.  There's a case to be made for removing this deletion logic
on the grounds that aclocal should know what it is doing, but for now
make it clear that we're deleting a file by saying so in the task log.

Signed-off-by: Ross Burton <ross.burton@arm.com>