uboot-extlinux-config.bbclass: add support for DTBOs
Also known as FDTO, DTO or Device Tree Overlay.

This adds a new property per label in extlinux.conf for which device
tree overlays to apply on top of the base device tree (the one specified
in FDT or guessed by the bootloader if only FDTDIR is set). It is a
space-separated list of paths to the .dtbo file(s).

Note that the base FDT needs to have symbols enabled (DTC_FLAGS needs to
have -@ in it) for an overlay to be applicable.

This is only applicable for non-FIT setups.

(From OE-Core rev: 7751337e71affb0e00528edc561e3404e80d4900)

Signed-off-by: Quentin Schulz <quentin.schulz@cherry.de>
Signed-off-by: Mathieu Dubois-Briand <mathieu.dubois-briand@bootlin.com>