rsync: clean up configure/configure.sh fiddling
The upstream Makefiles tell autoconf to write the generated script to
configure.sh instead of the idiomatic configure.  We now remove all of
the Makefile rules that refer to configure.sh (makefile-no-rebuild.patch)
but some pieces remained, so remove them too and delete the existing
configure.sh to avoid confusion for anyone looking at the build treee.

Also add a comment explaining why autotools-brokensep is used.

Signed-off-by: Ross Burton <ross.burton@arm.com>