rsync: clean up configure/configure.sh fiddling
rsync's Makefiles use autoconf to write configure.sh and there are
Makefile rules that depend on it, so we used to have to copy our
generated configure to configure.sh.  We now delete those rules to avoid
them getting in the way, so we can also remove the configure to
configure.sh fiddling and just delete configure.sh at the beginning to
avoid confusion for anyone looking at the build treee.

Also add a comment explaining why autotools-brokensep is used.

Signed-off-by: Ross Burton <ross.burton@arm.com>