conf/bitbake.conf: move addfragments to just after the local.conf include
There was a report that the existing addfragments statement comes in too late,
specifically after the DISTRO/MACHINE includes have already been handled. The goal
for fragments content is to behave exactly like the settings in local.conf would
and so we need to handle the fragments just after that file.

Signed-off-by: Alexander Kanavin <alex@linutronix.de>