kernel-fitimage: re-write its code in python
This is a refactoring without any functional change.

The goal is to split future proof parts of the kernel-fitimage in a back
and forward compatible way into a new more generic bbclass or into a
Python library which can be used by the legacy kernel-fitimage.bbclass
or any other recipe which generates a fitImage.

Re-writing the code line by line (as far as possible) in Python is only
the first step before the code becomes more extendable and more
maintainable and finally split-able.

kernel-uboot needs to be Python too. This allows to cleanup the
linux_comp handling and to simplify the call the function from other
Python functions.

Signed-off-by: Adrian Freihofer <adrian.freihofer@siemens.com>