meta-poky/conf/templates/default/bblayers.conf.sample: do not assume integrated poky repo
The list of layers was written assuming the template is used with
the 'integrated' poky checkout, where content of meta-yocto is
'overlaid' on content of openembedded-core, and the layers are
all in the same top level poky repository directory.

This doesn't work when oe-core and meta-yocto are cloned separately,
where relative path that starts from oe-core would have to go one
level above oe-core and then go into meta-yocto, assuming the
separate repo clones are in the same top level directory.

This requires a custom tweak to bblayers.conf when it makes its way from
meta-yocto into poky that would strip the ../meta-yocto/ part in
BBLAYERS.

This was prompted by
https://lists.openembedded.org/g/openembedded-architecture/topic/layer_layout_poky_vs/111010762

particularly the last email
https://lists.openembedded.org/g/openembedded-architecture/message/2129
indicating that such a tweak is possible.

Signed-off-by: Alexander Kanavin <alex@linutronix.de>