meta-poky/conf/templates: add a blank template
The default oe-core/poky templates have a number of settings
(such as distro and machine) and are thus unsuitable for
building up configurations from fragments. This adds a template
with an empty local.conf, so that it can be used as the starting
point for fragment-driven configuration.

There will also be a blank template in oe-core; the difference is
that this one includes poky layers and the oe-core one does not.

Signed-off-by: Alexander Kanavin <alex@linutronix.de>