SPDX: Add Product data
Adds SPDX "product" data which is an SPDX package that describes the
upstream source code, and things that our recipe statically (e.g.
without actually building anything) knows about it.

This includes the VEX data, many recipe variables. The generated runtime
packages are changed to no longer include the VEX data, but instead use
the "packagedBy" relationship to link them back to the product. This
significantly improves the speed of processing the VEX data, and also
allows it to be generated without actually building any source code

Signed-off-by: Joshua Watt <JPEWhacker@gmail.com>