rust: Reduce the unit tests that are ignored
We had previously ignored certain unit tests due to failures in earlier versions.
However, with the latest version of Rust, some of these unit tests are passing.
As a result, we are adding these unit tests back to execution.

The rust-1.85.1 has been successfully tested with these unit tests inclusions.

Signed-off-by: Deepesh Varatharajan <Deepesh.Varatharajan@windriver.com>
Signed-off-by: Harish Sadineni <Harish.Sadineni@windriver.com>