oe-selftest: devtool: split tap detection into function
Make the check for tap devices available as a function which can be used
by other tests as well.

Remove /etc/runqemu-nosudo check for tap device detection.
The /etc/runqemu-nosudo file is a state file created by the
runqemu-gen-tapdevs script. Tap device detection now relies only on the
presence of tap interfaces, making the check more robust and not
dependent on the runqemu-gen-tapdevs script, which is just one way of
creating the tap interfaces.

Simplify the code by removing the fallback to ifconfig if the ip command
is not available. ip commands are nowadays available on all host
machines. The transition from ifconfig to ip has taken place long time
ago e.g. for the runqemu-gen-tapdevs script.

Signed-off-by: Adrian Freihofer <adrian.freihofer@siemens.com>