rust: Enable Nightly builds
We have successfully completed the integration and build of the
Rust nightly toolchain within the Yocto environment. The following
BitBake commands are executing without issues: rust, cargo, and
nativesdk-rust. The build process has been validated using the
latest available nightly tarball as of today.

Based on the commit history, the stage0 compiler in the upstream
master branch is typically updated every 15 to 20 days. The current
stage0 compiler version in use is from 2025-05-26. Once a newer version
is released, we will revalidate the build process accordingly.

Until now, oe-selftest was not included in our nightly build validation.

Changes:

disable-rustfmt-check.patch
This patch disables the rustfmt check during the build process. The check
is specifically intended for nightly builds and is not applicable to beta
or stable releases. Additionally, it attempts to download the rustfmt
tarball during the build, which leads to failures. Since this check is
primarily used to ensure that new commits conform to formatting standards
upstream and is not necessary for our use case we have disabled it for
atleast for now.

From rust_1.87
Previously, we copied additional tools from the stage2-tools directory
to the bin directory. However, starting with Rust version 1.87, the
stage2-tools are no longer built as part of the standard build process.
For now we are copying the tools from the stage1-tools directory. Further
analysis is required to fully understand the implications of this change.

Limitations:
Our local patches are not rebased yet. So that will be rebased and
pushed soon to this branch.

Signed-off-by: Deepesh Varatharajan <Deepesh.Varatharajan@windriver.com>