beaglebone-yocto: remove redundant boot files
The beaglebone-yocto MACHINE uses wic to create the final image. The wks
file defines a boot partition like that:
  part /boot --source bootimg_partition --ondisk mmcblk0

This copies the files listed in IMAGE_BOOT_FILES variable from the
DEPLOY_DIR_IMAGE to the dedicated boot partition. Therefore installing
the kernel and the u-boot also into the rootfs is redundant.
This is tested on a beaglebone-black:
* Without this commit: systemd boots. Unmounting /boot shows the useless
  artifacts in the rootfs /boot directory.
* With this commit: The system boots. Unmounting /boot leads to an empty
  /boot directory.

Signed-off-by: Adrian Freihofer <adrian.freihofer@siemens.com>