beaglebone-yocto: default to beaglebone black
The Beaglebone Black is the newest and still available board. It is also
the recommended board on https://www.beagleboard.org/boards.

Booting from SDCard is possible with all three device trees. However,
for example, the internal memory of a Beaglebone Black requires the
correct DTB. This change makes this work by default, improving the user
experience for those who buy a new Yocto reference board to get
started.

Signed-off-by: Adrian Freihofer <adrian.freihofer@siemens.com>