beaglebone-yocto: default to beaglebone black
The Beaglebone black is the newest and still available board. It's also
the recommended board on https://www.beagleboard.org/boards.

Booting from SDCard is possible with all three device-trees. However,
for example the internal memory of a Beaglebone black requires the right
DTB. This change makes this working by default which improves the user
experience for users who buy a new Yocto reference board for getting
started.

Signed-off-by: Adrian Freihofer <adrian.freihofer@siemens.com>