bb/utils: improve cpu_count()
Move the implementation of cpu_count() from oe.utils here. This version
also uses sched_getaffinity() but also has arguments to control the
range of values, which is useful to cap the number of threads on systems
with hundreds of cores where the overhead of managing that many threads
can outweight the performance gains.

Signed-off-by: Ross Burton <ross.burton@arm.com>