bitbake: use bb.utils.cpu_count() instead of more naive counts
os.cpu_count() or multiprocessing.cpu_count() shouldn't be used, because
they return the number of physical cores on a system and not the number
of cores that the current process actually has access to (which can be
controlled via cgroups or taskset).

Note that this will implicitly cap the number of parse threads to 64,
but in my testing this is a minimal difference. On an Altra Max with 128
cores, it takes bitbake 31s to parse from scratch using all 128 cores, or
33s with just 64 cores.

Signed-off-by: Ross Burton <ross.burton@arm.com>