dbus: glib is only used in the test suite
The glib/gio libraries are only used in a few of the tests:

$ grep use_glib
meson.build:use_glib = glib.found() and gio.found()
test/meson.build:if use_glib
test/meson.build:if message_bus and tools and platform_unix and use_glib
test/meson.build:if use_glib
test/meson.build:if message_bus and tools and platform_unix and use_glib

So there's no need to depend on glib if we're not building the tests.
Explicitly enable/disable the modular tests option to ensure that when
the tests are not being installed we don't still need glib to build them.

Signed-off-by: Ross Burton <ross.burton@arm.com>