wic: add --keep-free option to set unused space
Currently, the content of the partition is filled by the filesystem
without leaving any unused free space. The --extra-space flag adds
extra space to the filesystem size, not to the partition.

Unused free space after the filesystem can be useful for some cases,
such as encrypting a partition at runtime. With --keep-free 32M, we
ensure that the last 32M of the partition is unused: this space does
not contain filesystem data and can store the LUKS2 header.

The implementation sets a difference between the partition and
filesystem size:
  - With --fixed-size, the keep-free space is removed from the
    filesystem size.
  - Otherwise (with or without --size flags), the keep-free space is
    added to the partition size.

(From OE-Core rev: b9c14c317d0389e74e7134d69baed7f2603352a4)

Signed-off-by: Pierre-Loup GOSSE <pierre-loup.gosse@smile.fr>
Signed-off-by: Mathieu Dubois-Briand <mathieu.dubois-briand@bootlin.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>