dbus-glib: remove unused packaging of tests
We're carrying a patch to install the tests, but we don't actually
install them by default or have any ptest infrastructure to run them.

As this is complicating the recipe for no reason, remove it all. If
someone in the future wants to run the tests they can do it with ptest
so they're exercised correctly.

(From OE-Core rev: 2f691058cdfb86e1ff6f5477b7a975dbd8283889)

Signed-off-by: Ross Burton <ross.burton@arm.com>
Signed-off-by: Mathieu Dubois-Briand <mathieu.dubois-briand@bootlin.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>