oe-selftest: devtool ide-sdk cover vscode remote debugging
This adds more tests to coverage for devtool ide-sdk, with VSCode. The
cmake test case has now a full remote debugging test on Qemu. The test
checks the generated launch.json and tasks.json files, starts
gdbserver and connects to it. The test verifies breakpoints, variables
and source file listing.

Signed-off-by: Adrian Freihofer <adrian.freihofer@siemens.com>