linux-firmware: skip ldflags sanity check
Some of the firmware blobs, for example qcom/apq8016/modem.mbn, are
actually ELF files. We don't want to hold these to our exacting standard
for link flags because we're not building them and they don't run on the
target directly.

Note that as this check parses the output of ${OBJDUMP} -p it behaves
differently with binutils vs llvm.  It looks like binutils bails early
as it doesn't know what the target architecture is, whereas LLVM is more
comprehensive in its support.

Signed-off-by: Ross Burton <ross.burton@arm.com>