cooker: add support for more than one server in BB_HASHSERVE_UPSTREAM
Making the use of python websockets mandatory when connecting to hashserve could
create several difficulties for users, especially when the python-python is not
available in the distribution used.
So it will be better to have a list of servers where for example we can
have a second fallback in case it is not possible to use the first one.

If it is not possible to use any server from the specified list, this will
cause a fatal error.

Signed-off-by: Jose Quaresma <jose.quaresma@oss.qualcomm.com>