bitbake: bitbake-setup: clarify that default answer to prompts is no
It is common practice to put the default choice in upper case for
yes/no interactive prompts, so that when people just hit enter, they
know what they are getting.  An example that linux users are probably
familiar with is "sensors-detect" from the "lm-sensors" package.

Unify all the prompts to be the same and indicate that the default
answer from hitting enter is a no with an upper case N.  No functional
changes.

(Bitbake rev: b927904eae17347d4567ea740cfe970b43510810)

Signed-off-by: Paul Gortmaker <paulg@kernel.org>
Signed-off-by: Mathieu Dubois-Briand <mathieu.dubois-briand@bootlin.com>